services/feature_jobs/batch/requirements.txt
numpy==1.26.4
pandas==2.1.0
psycopg2-binary==2.9.9
